fix: unauthorized user can't have ability to create contracts, only to log in
