+++ src/hooks/index.ts
@@ -10,3 +10,4 @@ export * from './useParticleNetwork';
 export * from './useBreakpoints';
 export * from './useAuthHandlers';
 export * from './useAdminPanel';
+export * from './useAuthConnectWallet';
